Moved the views file prior to breaking it apart.
