+++ kakao/src/main/kotlin/com/agoda/kakao/Extensions.kt
@@ -4,6 +4,10 @@ package com.agoda.kakao
 
 import com.agoda.kakao.builders.DataBuilder
 import com.agoda.kakao.builders.ViewBuilder
+import com.agoda.kakao.views.KEmptyAdapterItem
+import com.agoda.kakao.views.KEmptyRecyclerItem
+import com.agoda.kakao.views.KListView
+import com.agoda.kakao.views.KRecyclerView
 
 /**
  * Calls childAt() on your view with base child
